remove cyclical dev-dependency on core
