+++ datafusion/datasource/src/memory.rs
@@ -765,27 +765,16 @@ mod memory_source_tests {
 
 #[cfg(test)]
 mod tests {
+    use crate::tests::{aggr_test_schema, make_partition};
+
     use super::*;
-    use arrow::array::{ArrayRef, Int32Array};
+
     use datafusion_physical_plan::expressions::lit;
-    use std::collections::HashMap;
 
     use arrow::datatypes::{DataType, Field};
     use datafusion_common::assert_batches_eq;
     use datafusion_common::stats::{ColumnStatistics, Precision};
     use futures::StreamExt;
-
-    // Return a RecordBatch with a single Int32 array with values (0..sz) in a field named "i"
-    pub fn make_partition(sz: i32) -> RecordBatch {
-        let seq_start = 0;
-        let seq_end = sz;
-        let values = (seq_start..seq_end).collect::<Vec<_>>();
-        let schema = Arc::new(Schema::new(vec![Field::new("i", DataType::Int32, true)]));
-        let arr = Arc::new(Int32Array::from(values));
-        let arr = arr as ArrayRef;
-
-        RecordBatch::try_new(schema, vec![arr]).unwrap()
-    }
 
     #[tokio::test]
     async fn exec_with_limit() -> Result<()> {
@@ -811,29 +800,6 @@ mod tests {
         ];
         assert_batches_eq!(expected, &results);
         Ok(())
-    }
-
-    /// Get the schema for the aggregate_test_* csv files
-    pub fn aggr_test_schema() -> SchemaRef {
-        let mut f1 = Field::new("c1", DataType::Utf8, false);
-        f1.set_metadata(HashMap::from_iter(vec![("testing".into(), "test".into())]));
-        let schema = Schema::new(vec![
-            f1,
-            Field::new("c2", DataType::UInt32, false),
-            Field::new("c3", DataType::Int8, false),
-            Field::new("c4", DataType::Int16, false),
-            Field::new("c5", DataType::Int32, false),
-            Field::new("c6", DataType::Int64, false),
-            Field::new("c7", DataType::UInt8, false),
-            Field::new("c8", DataType::UInt16, false),
-            Field::new("c9", DataType::UInt32, false),
-            Field::new("c10", DataType::UInt64, false),
-            Field::new("c11", DataType::Float32, false),
-            Field::new("c12", DataType::Float64, false),
-            Field::new("c13", DataType::Utf8, false),
-        ]);
-
-        Arc::new(schema)
     }
 
     #[tokio::test]
